requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
pydantic
groq
chromadb
sentence-transformers
